Split file data store to subpackage and separate files
