+++ src/smif/data_layer/file/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""Data access modules for loading system-of-systems model configuration
+"""
+
+# import classes for access like ::
+#         from smif.data_layer.file import YamlConfigStore`
+from smif.data_layer.file.file_config_store import YamlConfigStore
+from smif.data_layer.file.file_data_store import CSVDataStore
+from smif.data_layer.file.file_metadata_store import FileMetadataStore
+
+# Define what should be imported as * ::
+#         from smif.data_layer.file import *
+__all__ = ['YamlConfigStore', 'CSVDataStore', 'FileMetadataStore']
